Added authorization control for 'postLike' API

+++ src/hooks/usePostDelete.tsx
@@ -1,3 +1,4 @@
+import { auth } from "@/firebase/clientApp";
 import { useState } from "react";
 
 export default function usePostDelete() {
@@ -7,25 +8,31 @@ export default function usePostDelete() {
     setPostDeletionLoading(true);
     console.log("Post Deletion is started");
 
-    const response = await fetch("/api/postDelete", {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ postDocPath: postDocPath }),
-    });
+    try {
+      const idToken = await auth.currentUser?.getIdToken();
 
-    if (!response.ok) {
-      if (response.status === 500) {
-        const { firebaseError } = await response.json();
-        console.error("Firebase Error while deleting post", firebaseError);
-      } else {
-        const { error } = await response.json();
-        console.error("Non-Firebase Error while deleting post", error);
+      if (!idToken) {
+        throw new Error("Id Token couldn't be get");
       }
-    } else {
-      console.log("Post successfully finished");
-      setPostDeletionLoading(false);
+
+      const response = await fetch("/api/postDelete", {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          authorization: `Bearer ${idToken}`,
+        },
+        body: JSON.stringify({ postDocPath: postDocPath }),
+      });
+
+      if (!response.ok) {
+        const { error } = await response.json();
+        throw new Error(error);
+      } else {
+        console.log("Post successfully finished");
+        setPostDeletionLoading(false);
+      }
+    } catch (error) {
+      console.error("Error while deleting post", error);
     }
   };
   return {
